Add CLI to follow README, update README folder name, remove unneeded var
